fixed all static code analysis issues
