+++ bundles/org.openhab.binding.surepetcare/src/main/java/org/openhab/binding/surepetcare/internal/discovery/SurePetcareDiscoveryService.java
@@ -30,10 +30,10 @@ import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.ThingUID;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerService;
-import org.openhab.binding.surepetcare.internal.data.SurePetcareDevice;
-import org.openhab.binding.surepetcare.internal.data.SurePetcareDevice.ProductType;
-import org.openhab.binding.surepetcare.internal.data.SurePetcareHousehold;
-import org.openhab.binding.surepetcare.internal.data.SurePetcarePet;
+import org.openhab.binding.surepetcare.internal.dto.SurePetcareDevice;
+import org.openhab.binding.surepetcare.internal.dto.SurePetcareHousehold;
+import org.openhab.binding.surepetcare.internal.dto.SurePetcarePet;
+import org.openhab.binding.surepetcare.internal.dto.SurePetcareDevice.ProductType;
 import org.openhab.binding.surepetcare.internal.handler.SurePetcareBridgeHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
