+++ bundles/org.openhab.binding.surepetcare/src/main/resources/ESH-INF/binding/binding.xml
