+++ bundles/org.openhab.binding.surepetcare/src/main/resources/ESH-INF/thing/bridge.xml
