+++ bundles/org.openhab.binding.surepetcare/src/test/java/org/openhab/binding/surepetcare/internal/data/SurePetcareTopologyTest.java
@@ -14,14 +14,17 @@ package org.openhab.binding.surepetcare.internal.data;
 
 import static org.junit.Assert.*;
 
+import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.junit.Test;
 import org.openhab.binding.surepetcare.internal.SurePetcareConstants;
+import org.openhab.binding.surepetcare.internal.dto.SurePetcareTopology;
 
 /**
  * The {@link SurePetcareTopologyTest} class implements unit test case for {@link SurePetcareTopology}
  *
  * @author Rene Scherer - Initial contribution
  */
+@NonNullByDefault
 public class SurePetcareTopologyTest {
 
     @Test
